+++ pepr.ts
@@ -9,14 +9,12 @@ import cfg from "./package.json";
 
 import { Component, setupLogger } from "./src/pepr/logger";
 import { operator } from "./src/pepr/operator";
-import {
-  setupAuthserviceSecret,
-  setupKeycloakClientSecret,
-} from "./src/pepr/operator/controllers/keycloak/authservice/config";
+import { setupAuthserviceSecret } from "./src/pepr/operator/controllers/keycloak/authservice/config";
 import { registerCRDs } from "./src/pepr/operator/crd/register";
 import { patches } from "./src/pepr/patches";
 import { policies, startExemptionWatch } from "./src/pepr/policies";
 import { prometheus } from "./src/pepr/prometheus";
+import { setupKeycloakClientSecret } from "./src/pepr/operator/controllers/keycloak/config";
 
 const log = setupLogger(Component.STARTUP);
